Manage S3 bucket by terraform
Terraform handles the whole lifecycle of the resources. So the
destroy phase must be handled.
With CloudFormation the S3 bucket cannot be deleted if not empty. But
the terraform aws provider has an option to force it:
https://www.terraform.io/docs/providers/aws/r/s3_bucket.html#force_destroy

So the S3 bucket was extracted from the CloudFormation and is now
managed from terraform. This way we can force terraform to destroy all
objects inside the S3 bucket and then destroy the bucket. So the
whole lifecycle is respected.

To be able to do this, I patched the CloudFormation template:
https://s3.amazonaws.com/epsagon-cloudformation/auto_template.json

You can find below the diff of the changes I applied.
I added a new variable `ExternalBucketName` that will be filled by the
name of the S3 bucket created by terraform. The S3 bucket is created
with the same expiration.
Then, with this variable, I added a condition to know if the bucket is
managed by CloudFormation or must be provided: `InternalBucket`.
The `EpsagonTrailBucket` CloudFormation resource is created only if
`InternalBucket` is true, so if `ExternalBucketName` is an empty string.
Then some conditions are added to use the internal CloudFormation
reference or the external name of the bucket, depending on the
situation.
All other resources stay the same.

If this patch is applied on your file on S3, this Terraform module will
be able to create the S3 bucket then run the CloudFormation and the destroy
phase will be also respected.

```diff
diff --git a/auto_template.json.orig b/auto_template.json
index 284d6f8..100c594 100644
--- a/auto_template.json.orig
+++ b/auto_template.json
@@ -35,6 +35,7 @@
     },
     "EpsagonTrailBucket": {
       "Type": "AWS::S3::Bucket",
+      "Condition": "InternalBucket",
       "Properties": {
         "LifecycleConfiguration": {
           "Rules": [
@@ -57,7 +58,9 @@
       "Type": "AWS::S3::BucketPolicy",
       "Properties": {
         "Bucket": {
-          "Ref": "EpsagonTrailBucket"
+          "Fn::If": ["InternalBucket",
+            { "Ref": "EpsagonTrailBucket" },
+            { "Fn::Sub": "${ExternalBucketName}" }]
         },
         "PolicyDocument": {
           "Version": "2012-10-17",
@@ -69,7 +72,9 @@
               },
               "Action": "s3:GetBucket*",
               "Resource": {
-                "Fn::Sub": "arn:aws:s3:::${EpsagonTrailBucket}"
+                "Fn::If": ["InternalBucket",
+                  { "Fn::Sub": "arn:aws:s3:::${EpsagonTrailBucket}" },
+                  { "Fn::Sub": "arn:aws:s3:::${ExternalBucketName}" }]
               }
             },
             {
@@ -79,7 +84,9 @@
               },
               "Action": "s3:PutObject",
               "Resource": {
-                "Fn::Sub": "arn:aws:s3:::${EpsagonTrailBucket}/AWSLogs/${AWS::AccountId}/*"
+                "Fn::If": ["InternalBucket",
+                  { "Fn::Sub": "arn:aws:s3:::${EpsagonTrailBucket}/AWSLogs/${AWS::AccountId}/*" },
+                  { "Fn::Sub": "arn:aws:s3:::${ExternalBucketName}/AWSLogs/${AWS::AccountId}/*" }]
               }
             }
           ]
@@ -150,7 +154,9 @@
           ]
         },
         "S3BucketName": {
-          "Ref": "EpsagonTrailBucket"
+          "Fn::If": ["InternalBucket",
+            { "Ref": "EpsagonTrailBucket" },
+            { "Fn::Sub": "${ExternalBucketName}" }]
         },
         "TrailName": "EpsagonMonitoringTrail",
         "EventSelectors": [
@@ -336,6 +342,14 @@
     "EpsagonSns": {
       "Description": "The Epsagon SNS Name for CF callback.",
       "Type": "String"
+    },
+    "ExternalBucketName": {
+      "Description": "Name of the external bucket. If empty, the bucket will be managed by CloudFormation.",
+      "Type": "String",
+      "Default": ""
     }
+  },
+  "Conditions": {
+    "InternalBucket": { "Fn::Equals": [{ "Fn::Sub": "${ExternalBucketName}" }, ""] }
   }
 }
```